+++ test/suite/previewManager.test.ts
@@ -32,7 +32,10 @@ suite("PreviewManager and pythonPanelPreview Tests", () => {
                 previewManager.startArepl().then((previewPanel)=>{
                     panel = previewPanel
                     console.log("preview panel started")
-                    done()
+                    // wait for default import to be inserted
+                    setTimeout(() => {
+                        done()
+                    }, 100);
                 }).catch((err)=>done(err))
             })
         })
@@ -53,8 +56,8 @@ suite("PreviewManager and pythonPanelPreview Tests", () => {
             setTimeout(()=>{
                 assert.equal(panel.webview.html.includes(`"x":3424523`), true, panel.webview.html)
                 done()
-            },4000)
-        })
+            },3000)
+        }, done)
     });
 
     suiteTeardown(function(){